% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.plus.R
\name{assess}
\alias{assess}
\alias{assess.plus}
\title{Assessment of performance metrics in plus models}
\usage{
assess(object, newx = NULL, newy, weights = NULL, ...)

\method{assess}{plus}(object, newx = NULL, newy, weights = NULL, ...)
}
\arguments{
\item{object}{A plus object}

\item{newx}{A matrix with new data to test performance}

\item{newy}{A vector with the observed classes to test performance}

\item{weights}{optional vector of case weights}

\item{\dots}{additional arguments}
}
\value{
A list with performance metrics for plus models
}
\description{
Assessment of performance metrics in plus models
}
\examples{
#' data(binexample)
x = binexample$x
y = binexample$y
s <- sample(seq(nrow(x)), 65, replace = FALSE)
x_train <- x[s, ]
x_test  <- x[-s, ]
y_train <- y[s]
y_test  <- y[-s]
fit <- plus(x_train, y_train)
assess(fit, newx = x_test, newy = y_test)
}
