% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_plus.R
\name{assess_plus}
\alias{assess_plus}
\title{Assesment of performance metrics in plus models}
\usage{
assess_plus(object, newx = NULL, newy, weights = NULL, ...)
}
\arguments{
\item{object}{A plus object}

\item{newx}{A matrix with new data to test performance}

\item{newy}{A vector with the observed classes to test performance}

\item{weights}{optional vector of case weights}

\item{\dots}{additional arguments}
}
\value{
A list with performance metrics for plus models
}
\description{
Assesment of performance metrics in plus models
}
\examples{
x_train <- matrix(rnorm(650 * 20), 650, 20)
x_test  <- matrix(rnorm(350 * 20), 350, 20)
y_train <- ifelse(rnorm(650) > 0, 1, 0)
y_test  <- ifelse(rnorm(350) > 0, 1, 0)
fit <- plus(x_train, y_train)
assess_plus(fit, newx = x_test, newy = y_test)
}
