% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Gets the Area under ROC curve usingthe concordance statistics}
\usage{
auc(y, prob, w)
}
\arguments{
\item{y}{A vector of binary clases}

\item{prob}{A vector of probabilities (logistic regression output)}

\item{w}{optional vector of case weights}
}
\value{
The AUC value for a plus object
}
\description{
Gets the Area under ROC curve usingthe concordance statistics
}
\examples{
x_train <- matrix(rnorm(650 * 20), 650, 20)
x_test  <- matrix(rnorm(350 * 20), 350, 20)
y_train <- ifelse(rnorm(650) > 0, 1, 0)
y_test  <- ifelse(rnorm(350) > 0, 1, 0)
fit <- plus(x_train, y_train)
p_response <- predict(fit, newx = x_test)
p_class <- plus:::predict.plus(fit, newx = x_test, type = "class")
auc(p_class, p_response)
}
