% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binexample.R
\docType{data}
\name{binexample}
\alias{binexample}
\title{List with 100 random values for binary classfication It contains a list
with x and y values. x.has 100 instances and 30 features. y is a vector
of 100 binary (1 and 0) values.
This dataset is for test plus classification.}
\format{
An object of class \code{"list"};
}
\usage{
data(binexample)
}
\description{
List with 100 random values for binary classfication It contains a list
with x and y values. x.has 100 instances and 30 features. y is a vector
of 100 binary (1 and 0) values.
This dataset is for test plus classification.
}
\examples{
data(binexample)
}
\keyword{datasets}
