% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_example.R
\docType{data}
\name{cutoff_example}
\alias{cutoff_example}
\title{Example of cutoff threshold from plus model for use in the consturctor function new_plus}
\format{
An object of class \code{"numeric"}; it has a numeric value of the model threshold.
}
\usage{
data(cutoff_example)
}
\description{
Example of cutoff threshold from plus model for use in the consturctor function new_plus
}
\examples{
data(cutoff_example)
}
\keyword{datasets}
