% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{PLUS package built-in example data.}
\format{A list whose length is 3:
\describe{
  \item{train_data}{Training data}
  \item{Label.obs}{Label object}
  \item{Label.true}{Ground true label}
}}
\usage{
example_data
}
\description{
PLUS package built-in example data.
}
\keyword{datasets}
