% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_plus_example.R
\docType{data}
\name{fit_plus_example}
\alias{fit_plus_example}
\title{Example of plus model for use in the consturctor function new_plus}
\format{
An object of class \code{"cv.glmnet"};
}
\usage{
data(fit_plus_example)
}
\description{
Example of plus model for use in the consturctor function new_plus
}
\examples{
data(fit_plus_example)
}
\keyword{datasets}
