% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auc.R
\name{get_auc}
\alias{get_auc}
\alias{get_auc.plus}
\title{Assessment of auc metrics in plus models}
\usage{
get_auc(object, newx = NULL, newy = NULL, weights = NULL, ...)

\method{get_auc}{plus}(object, newx = NULL, newy = NULL, weights = NULL, ...)
}
\arguments{
\item{object}{A plus object}

\item{newx}{A matrix with new data to test performance}

\item{newy}{A vector with the observed classes to test performance}

\item{weights}{optional vector of case weights}

\item{\dots}{additional arguments}
}
\description{
Assessment of auc metrics in plus models
}
\examples{
data(binexample)
x <- binexample$x
y <- binexample$y
train <- sample(seq(length(y)), 65, replace = FALSE)
x_train <- x[train, ]
x_test  <- x[-train, ]
y_train <- y[train]
y_test  <- y[-train]
fit <- plus(x_train, y_train)
get_auc(fit, x_test, y_test)
}
