% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Get predictions to asses model performance from a fitted plus model.}
\usage{
get_predictions(object, newx, newy, plus = TRUE)
}
\arguments{
\item{object}{A plus object with glmnet classification}

\item{newx}{New data to predict the class}

\item{newy}{The observed class to meassure the performance of the model}

\item{plus}{Logical. If it is true use the cutoff threshold from the plus model.
This tends to reduce the false negatives and increase the recall. Othewise
gets the predicted value from glmnet default predicted response.#'}
}
\value{
A \code{data.frame} containing four columns. The truth independent
value (observed y), the probability to belong in Class1, the probability to
belong Class2 and the predicted class either the plus (use cutoff threshold)
or glmnet model.
}
\description{
Get predictions to asses model performance from a fitted plus model.
}
\examples{
data(binexample)
x = binexample$x
y = binexample$y
s <- sample(seq(nrow(x)), 65, replace = FALSE)
train_x <- x[s, ]
train_y <- y[s]
newx <- x[-s, ]
newy <- y[-s]
model <- plus(train_x, train_y)
get_predictions(model, newx, newy)
get_predictions(model, newx, newy, plus = FALSE)
}
