% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacs.R
\docType{data}
\name{lacs}
\alias{lacs}
\title{Data frame with 5990 abstracts. Each abstract belongs to
one of both classes, positive and unknown. Abstracts from parasite class
are from ZOVER and GMPD database. Abstracs from unknown class are random
abstracts retrived from crossref.}
\format{
An object of class \code{"data.frame"};
}
\usage{
data(lacs)
}
\description{
Data frame with 5990 abstracts. Each abstract belongs to
one of both classes, positive and unknown. Abstracts from parasite class
are from ZOVER and GMPD database. Abstracs from unknown class are random
abstracts retrived from crossref.
}
\examples{
data(lacs)
}
\keyword{datasets}
