% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plus.R
\name{new_plus}
\alias{new_plus}
\title{Constructor function for the class plus
A constructor for the plus class}
\usage{
new_plus(
  fit_plus = list(),
  pred_y = matrix(),
  cutoff = numeric(),
  predicted_coefficients = Matrix::Matrix(),
  call = character()
)
}
\arguments{
\item{fit_plus}{a cv.glmnet object. Is the output of the plus
algorithm performance}

\item{pred_y}{A vector of predictied probabilities of the plus model}

\item{cutoff}{The threshold cut of the plus model}

\item{predicted_coefficients}{the coefficientes values predicted for the plus model}

\item{call}{Call this function}
}
\value{
an object of class plus
}
\description{
Constructor function for the class plus
A constructor for the plus class
}
\examples{
new_plus(fit_plus = fit_plus_example,
pred_y = predicted_y_example,
cutoff = cutoff_example,
predicted_coefficients = predicted_coefficients_example,
call = match.call()
)
}
