% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLUS.R
\name{PLUS}
\alias{PLUS}
\title{PLUS main function.}
\usage{
PLUS(train_data = train_data, Label.obs = Label.obs,
  Sample_use_time = 30, l.rate = 1, qq = 0.1)
}
\arguments{
\item{train_data}{training data.}

\item{Label.obs}{.Please write what this parameter means}

\item{Sample_use_time}{Please write what this parameter means.}

\item{l.rate}{Learning rate.}

\item{qq}{Please write what qq means.}
}
\value{
A list consist of three objects including predicted y, predicted coefficient, cutoff.
}
\description{
PLUS main function.
}
