% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_cv_lognet.R
\name{plus_cv_lognet}
\alias{plus_cv_lognet}
\title{Evaluation metric of plus classification}
\usage{
plus_cv_lognet(
  predmat,
  y,
  type.measure = c("deviance", "class", "auc", "mse", "mae"),
  weights = NULL
)
}
\arguments{
\item{predmat}{A prediction matrix given the plus model (lambda min)}

\item{y}{A vector of a binary virable to test performance}

\item{type.measure}{The type of performance metric ("deviance", "class", "auc", "mse", "mae")}

\item{weights}{optional vector of case weights}
}
\value{
If the type.measure is AUC returns auc value. Otherwise returna vector of values given the metrics.
}
\description{
Evaluation metric of plus classification
}
\examples{
x_train <- matrix(rnorm(650 * 20), 650, 20)
x_test  <- matrix(rnorm(350 * 20), 350, 20)
y_train <- ifelse(rnorm(650) > 0, 1, 0)
y_test  <- ifelse(rnorm(350) > 0, 1, 0)
fit <- plus(x_train, y_train)
p <- predict(fit,newx = x_test)
plus_cv_lognet(p, y_test, type.measure = "auc")
}
