% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_object_example.R
\docType{data}
\name{plus_object_example}
\alias{plus_object_example}
\title{Example of plus object from new_plus constructor}
\format{
An object of class \code{"plus"};
}
\usage{
data(plus_object_example)
}
\description{
Example of plus object from new_plus constructor
}
\examples{
data(plus_object_example)
}
\keyword{datasets}
