% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_coefficients_example.R
\docType{data}
\name{predicted_coefficients_example}
\alias{predicted_coefficients_example}
\title{Example of prediction from plus model for use in the consturctor function new_plus}
\format{
An object of class \code{"dgCMatrix"}; it contains the coefficients of the model for each variable in the rows
}
\usage{
data(predicted_coefficients_example)
}
\description{
Example of prediction from plus model for use in the consturctor function new_plus
}
\examples{
data(predicted_coefficients_example)
}
\keyword{datasets}
