% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_y_example.R
\docType{data}
\name{predicted_y_example}
\alias{predicted_y_example}
\title{Example of prediction from plus model for use in the consturctor function new_plus}
\format{
An object of class \code{"matrix"}; it has one dimenssion an n = 1000 rows
}
\usage{
data(predicted_y_example)
}
\description{
Example of prediction from plus model for use in the consturctor function new_plus
}
\examples{
data(predicted_y_example)
}
\keyword{datasets}
