% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_plus.R
\name{validate_plus}
\alias{validate_plus}
\title{validate plus_object constructed via new_plus function}
\usage{
validate_plus(plus_object)
}
\arguments{
\item{plus_object}{A plus object constructed with new_plus function}
}
\value{
A validated plus_object
}
\description{
validate plus_object constructed via new_plus function
}
\examples{
validate_plus(
new_plus(fit_plus = fit_plus_example,
pred_y = predicted_y_example,
cutoff = cutoff_example,
predicted_coefficients = predicted_coefficients_example)
)
}
