% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{plus}
\alias{plus}
\title{plus model helper function}
\usage{
plus(x = x, y = y, sample_use_time = 30, learning_rate = 1, qq = 0.1)
}
\arguments{
\item{x}{A set of independent numeric variables.}

\item{y}{A binary dependent variable}

\item{sample_use_time}{Number of use time}

\item{learning_rate}{learning rate}

\item{qq}{quantile threshold}
}
\value{
and object of the class plus
}
\description{
plus model helper function
}
\examples{
x = matrix(rnorm(1000 * 5), 1000, 5)
y = ifelse(rnorm(1000) > 0, 1, 0)
plus(x, y)
}
