% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.plus.R
\name{print.plus}
\alias{print.plus}
\title{print a plus object}
\usage{
\method{print}{plus}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fitted plus object}

\item{digits}{significant digits in printout}

\item{\dots}{additional print arguments}
}
\value{
information about the plus model
}
\description{
print a plus object
}
\examples{
x <- matrix(rnorm(1000 * 20), 1000, 20)
y <- ifelse(rnorm(1000) > 0, 1, 0)
fit <- plus(x, y)
print(fit)
}
\keyword{models}
\keyword{regression}
